Fix semantic

+++ src/parsing/parsing.c
@@ -6,7 +6,7 @@
 /*   By: mcogne-- <mcogne--@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/14 22:38:55 by mcogne--          #+#    #+#             */
-/*   Updated: 2024/12/14 23:12:35 by mcogne--         ###   ########.fr       */
+/*   Updated: 2024/12/15 22:16:02 by mcogne--         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,9 @@ short	parsing(t_minishell *env)
 	if (get_input(env))
 		return (1);
 	debug_print_input(env->input);
+	if (analys_semantic(env))
+		return (1);
+	debug_print_commands(env->commands);
 	delete_input(env);
 	return (0);
 }
